implement preview peeking beyond current context
This can help the user decide whether including more context might be
useful.
Since the dictionary and context window block a large portion of
the screen, the user previously often had to press one of the buttons to
expand the context to take a look (and undo the operation if this was
not the case).
The preview feature will also be helpful when further improving the
behaviour of the smart context selector.
The length of the context preview can be adjusted with the setting
`custom_context_peek_length`.