implement new algorithm for delimiter tracking
The new algorithm uses a table to save the location of all delimiters
within the required range of the context.
Multiple consecutive delimiting characters at treated like a single
delimiter, and a label is applied to differentiate sentence delimiters
from part-of-sentence delimiters.

A second step then uses the stable to look up a requested location in
(s, p) format, resulting in a range of potential delimiting characters
(if multiple consecutive are available).

In a third step additional adjustments are applied that define which of
the leading / trailing delimiting characters should be included and
which ought to be removed.
This step especially includes a parsing run through any part delimiters
in the context that ensures matching delimiters are included
automatically (and any non-matching removed respectively).

The presence of the categorized delimiter table also is a major step
towards the implementation of truly position-aware buttons in the custom
context menu.

With this commit the first and second steps are complete, the additional
adjustments are still a work in progress at this point.

Minor other adjustments:
* fix mangled string format options (accidental search-and-replace?)
* part-of-sentence delimiter characters do not include the characters
  from the sentence delimiter setting anymore.
* Add lower limit of zero sentences for the << >> buttons