simplify delim handling by removing "matched_word"
Doing so fixes a bug were one delimiter too much is selected if there is
a delimiter immediately adjacent to the matched word.
Also, the presence of a delimiter introduced more complexity than it was
worth, and the original main reason to have it in there is long since
gone (nested sentence->part delimiter table requiring a first element).