restructure context / delimiter tables
The tables are now structured as follows:
  `context_table.prev` /   `context_table.next`
`delimiter_table.prev` / `delimiter_table.next`
Previously it was:
  `previous_context_table` / `next_context_table`
`previous_delimiter_table` / `next_delimiter_table`
This allows us to indicate which context is meant with a single
indicator "prev" / "next" that is shared by the context and the
delimiter table. It also avoids using local variable names as indicator
as was the case before.