implement consistent buttons via pos. optimization
Background - the current button behaviour for adjusting parts/sentences
fails in some cases:
* If the char or part offset is larger than a whole sentence.
  Example: Starting at 0,5,0, incrementing by one sentence will result
  in 1,0,0, which will reduce (not increase) the length for sentences
  that are less than 5 parts long (in case of a sentence with 5 parts no
  change will occur).
* If there is a mix of positive and negative offsets, we do not know
  what to do on a sentence increment, since we do not know whether c+p
  moves us to the left or the right of the current sentence offset (or
  results in a net offset of zero).

To eliminate the aforementioned issues, the button logic now optimizes
the position after each update.
The optimized position guarantees that a char offset is always shorter
than a sentence part and a part offset always shorter than a whole
sentence.  It also guarantees that all position values are positive
(except for small adjustments as to which delimiting characters are
displayed / removed).

Additional changes:
- Ensure consistent button behavior for part and sentence increment with
  regard to reset/increment.
- Treat whitespace as part of a delimiter - this removes the need to
  apply whitespace trim adjustments afterwards.
- Introduce "whitespace" category in delimiter table (whitespace that is
  not part of a delimiter).